requirements.txt
livekit-agents
livekit-plugins-deepgram
livekit-plugins-silero
python-dotenv
livekit-plugins-openai